+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Added
 - Stats for every server
 - Bold selected server stat
+- Games won stat to selection menu on server stats
 
 # Changelog (Alpha Versions)
 ## [Release Alpha 1.5](https://github.com/Nevah5/wordlebot/releases/tag/a1.5) - 16.02.2022